Update to selectors and nested rules

+++ src/main/java/org/fernice/flare/font/WritingMode.kt
@@ -10,14 +10,14 @@ import kotlin.experimental.and
 class WritingMode(private val flags: Byte) {
 
     companion object {
-        const val RTL: Byte = 1 shl 0
+        const val RTL: Byte = (1 shl 0).toByte()
 
-        const val VERTICAL: Byte = 1 shl 1
-        const val VERTICAL_LR: Byte = 1 shl 2
+        const val VERTICAL: Byte = (1 shl 1).toByte()
+        const val VERTICAL_LR: Byte = (1 shl 2).toByte()
 
-        const val LINE_INVERTED: Byte = 1 shl 3
-        const val SIDEWAYS: Byte = 1 shl 4
-        const val UPRIGHT: Byte = 1 shl 5
+        const val LINE_INVERTED: Byte = (1 shl 3).toByte()
+        const val SIDEWAYS: Byte = (1 shl 4).toByte()
+        const val UPRIGHT: Byte = (1 shl 5).toByte()
 
         private const val ZERO: Byte = 0
     }
